+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 - Progress bar now increments correctly
+- Drive selection now changes config selection back to "none" when selecting drive that doesn't have a config file
 
 ### Changed
 - Backup run function now breaks immediately when aborted